+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### New Features
 
-- Add scheduler address in configuration instead of reading from Scheduler `/workers/config` endpoint. (#606)
+- Add workerpool address in configuration instead of reading from Scheduler `/workers/config` endpoint. (#607)
 
 ## [[8.5.0]](https://github.com/iExecBlockchainComputing/iexec-worker/releases/tag/v8.5.0) 2024-06-19